+++ Blueprints/src/BlueprintSchema.php
@@ -10,39 +10,25 @@ namespace RocketTheme\Toolbox\Blueprints;
  */
 class BlueprintSchema
 {
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $items = [];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $rules = [];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $nested = [];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $dynamic = [];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $filter = ['validation' => true];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $ignoreFormKeys = ['fields' => 1];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     protected $types = [];
 
     /**
@@ -52,7 +38,7 @@ class BlueprintSchema
      */
     public function __construct($serialized = null)
     {
-        if (is_array($serialized) && !empty($serialized)) {
+        if (\is_array($serialized) && !empty($serialized)) {
             $this->items = (array) $serialized['items'];
             $this->rules = (array) $serialized['rules'];
             $this->nested = (array) $serialized['nested'];
@@ -92,6 +78,7 @@ class BlueprintSchema
     {
         foreach ($this->dynamic as $key => $data) {
             $field = &$this->items[$key];
+
             foreach ($data as $property => $call) {
                 $action = 'dynamic' . ucfirst($call['action']);
 
@@ -127,7 +114,7 @@ class BlueprintSchema
      */
     public function get($name, $default = null, $separator = '.')
     {
-        $name = $separator !== '.' ? strtr($name, $separator, '.') : $name;
+        $name = $separator !== '.' ? str_replace($separator, '.', $name) : $name;
 
         return isset($this->items[$name]) ? $this->items[$name] : $default;
     }
@@ -143,7 +130,7 @@ class BlueprintSchema
      */
     public function set($name, $value, $separator = '.')
     {
-        $name = $separator !== '.' ? strtr($name, $separator, '.') : $name;
+        $name = $separator !== '.' ? str_replace($separator, '.', $name) : $name;
 
         $this->items[$name] = $value;
         $this->addProperty($name);
@@ -215,13 +202,14 @@ class BlueprintSchema
             $this->rules = array_merge($this->rules, $value['rules']);
         }
 
-        $name = $separator !== '.' ? strtr($name, $separator, '.') : $name;
+        $name = $separator !== '.' ? str_replace($separator, '.', $name) : $name;
 
         if (isset($value['form'])) {
             $form = array_diff_key($value['form'], ['fields' => 1, 'field' => 1]);
         } else {
             $form = [];
         }
+
         $items = isset($this->items[$name]) ? $this->items[$name] : ['type' => '_root', 'form_field' => false];
 
         $this->items[$name] = $items;
@@ -230,11 +218,13 @@ class BlueprintSchema
         $prefix = $name ? $name . '.' : '';
         $params = array_intersect_key($form, $this->filter);
         $location = [$name];
+
         if (isset($value['form']['field'])) {
             $this->parseFormField($name, $value['form']['field'], $params, $prefix, '', $merge, $location);
         } elseif (isset($value['form']['fields'])) {
             $this->parseFormFields($value['form']['fields'], $params, $prefix, '', $merge, $location);
         }
+
         $this->items[$name] += ['form' => $form];
 
         return $this;
@@ -253,7 +243,7 @@ class BlueprintSchema
     {
         $nested = $this->getNested($name, $separator);
 
-        if (!is_array($nested)) {
+        if (!\is_array($nested)) {
             $nested = [];
         }
 
@@ -343,7 +333,7 @@ class BlueprintSchema
      */
     protected function getPropertyRecursion($property, $nested)
     {
-        if (empty($nested) || !is_array($nested) || !isset($property['type'])) {
+        if (empty($nested) || !\is_array($nested) || !isset($property['type'])) {
             return $property;
         }
 
@@ -352,12 +342,13 @@ class BlueprintSchema
                 if ($key === '') {
                     continue;
                 }
-                $name = is_array($value) ? $key : $value;
+
+                $name = \is_array($value) ? $key : $value;
                 $property['fields'][$key] = $this->getPropertyRecursion($this->get($name), $value);
             }
         } elseif ($property['type'] === '_parent' || !empty($property['array'])) {
             foreach ($nested as $key => $value) {
-                $name = is_array($value) ? "{$property['name']}.{$key}" : $value;
+                $name = \is_array($value) ? "{$property['name']}.{$key}" : $value;
                 $property['fields'][$key] = $this->getPropertyRecursion($this->get($name), $value);
             }
         }
@@ -408,7 +399,8 @@ class BlueprintSchema
                 // TODO: Add support for adding defaults to collections.
                 continue;
             }
-            if (is_array($value)) {
+
+            if (\is_array($value)) {
                 // Recursively fetch the items.
                 $list = $this->buildDefaults($value);
 
@@ -441,10 +433,10 @@ class BlueprintSchema
     {
         foreach ($data2 as $key => $field) {
             $val = isset($rules[$key]) ? $rules[$key] : null;
-            $rule = is_string($val) ? $this->items[$val] : null;
+            $rule = \is_string($val) ? $this->items[$val] : null;
 
-            if ((array_key_exists($key, $data1) && is_array($data1[$key]) && is_array($field) && is_array($val) && !isset($val['*']))
-                || (!empty($rule['type']) && $rule['type'][0] === '_')) {
+            if ((array_key_exists($key, $data1) && \is_array($data1[$key]) && \is_array($field) && \is_array($val) && !isset($val['*']))
+                || (!empty($rule['type']) && strpos($rule['type'], '_') === 0)) {
                 // Array has been defined in blueprints and is not a collection of items.
                 $data1[$key] = $this->mergeArrays($data1[$key], $field, $val);
             } else {
@@ -468,7 +460,7 @@ class BlueprintSchema
      */
     protected function parseFormFields(array $fields, array $params, $prefix = '', $parent = '', $merge = false, array $formPath = [])
     {
-        if (isset($fields['type']) && !is_array($fields['type'])) {
+        if (isset($fields['type']) && !\is_array($fields['type'])) {
             return;
         }
 
@@ -490,7 +482,7 @@ class BlueprintSchema
     protected function parseFormField($key, array $field, array $params, $prefix = '', $parent = '', $merge = false, array $formPath = [])
     {
         // Skip illegal field (needs to be an array).
-        if (!is_array($field)) {
+        if (!\is_array($field)) {
             return;
         }
 
@@ -514,12 +506,13 @@ class BlueprintSchema
 
         $isInputField = !isset($properties['input@']) || $properties['input@'];
 
+        $propertyExists = isset($this->items[$key]);
         if (!$isInputField) {
             // Remove property if it exists.
-            if (isset($this->items[$key])) {
+            if ($propertyExists) {
                 $this->removeProperty($key);
             }
-        } elseif (!isset($this->items[$key])) {
+        } elseif (!$propertyExists) {
             // Add missing property.
             $this->addProperty($key);
         }
@@ -535,6 +528,7 @@ class BlueprintSchema
                 $path = explode('.', $key);
                 array_pop($path);
                 $parent = '';
+
                 foreach ($path as $part) {
                     $parent .= ($parent ? '.' : '') . $part;
                     if (!isset($this->items[$parent])) {
@@ -556,7 +550,7 @@ class BlueprintSchema
     protected function getFieldKey($key, $prefix, $parent)
     {
         // Set name from the array key.
-        if ($key && $key[0] === '.') {
+        if (strpos($key[0], '.') === 0) {
             return ($parent ?: rtrim($prefix, '.')) . $key;
         }
 
@@ -572,7 +566,7 @@ class BlueprintSchema
         }
 
         foreach ($properties as $name => $value) {
-            if (!empty($name) && ($name[0] === '@' || $name[strlen($name) - 1] === '@')) {
+            if (strpos($name[0], '@') !== false) {
                 $list = explode('-', trim($name, '@'), 2);
                 $action = array_shift($list);
                 $property = array_shift($list);
@@ -600,9 +594,10 @@ class BlueprintSchema
 
         $nested = &$this->nested;
         foreach ($parts as $part) {
-            if (!isset($nested[$part]) || !is_array($nested[$part])) {
+            if (!isset($nested[$part]) || !\is_array($nested[$part])) {
                 $nested[$part] = [];
             }
+
             $nested = &$nested[$part];
         }
 
@@ -624,9 +619,10 @@ class BlueprintSchema
 
         $nested = &$this->nested;
         foreach ($parts as $part) {
-            if (!isset($nested[$part]) || !is_array($nested[$part])) {
+            if (!isset($nested[$part]) || !\is_array($nested[$part])) {
                 return;
             }
+
             $nested = &$nested[$part];
         }
 
@@ -642,7 +638,7 @@ class BlueprintSchema
      */
     protected function getRule($rule)
     {
-        if (isset($this->rules[$rule]) && is_array($this->rules[$rule])) {
+        if (isset($this->rules[$rule]) && \is_array($this->rules[$rule])) {
             return $this->rules[$rule];
         }
         return [];
@@ -661,11 +657,11 @@ class BlueprintSchema
 
         foreach ($data as $key => $field) {
             $val = isset($rules[$key]) ? $rules[$key] : (isset($rules['*']) ? $rules['*'] : null);
-            $rule = is_string($val) ? $this->items[$val] : null;
+            $rule = \is_string($val) ? $this->items[$val] : null;
 
             if ($rule || isset($val['*'])) {
                 // Item has been defined in blueprints.
-            } elseif (is_array($field) && is_array($val)) {
+            } elseif (\is_array($field) && \is_array($val)) {
                 // Array has been defined in blueprints.
                 $array += $this->extraArray($field, $val, $prefix . $key . '.');
             } else {
@@ -685,7 +681,7 @@ class BlueprintSchema
     {
         $params = $call['params'];
 
-        if (is_array($params)) {
+        if (\is_array($params)) {
             $function = array_shift($params);
         } else {
             $function = $params;
@@ -695,19 +691,20 @@ class BlueprintSchema
         $list = explode('::', $function, 2);
         $f = array_pop($list);
         $o = array_pop($list);
+
         if (!$o) {
-            if (function_exists($f)) {
-                $data = call_user_func_array($f, $params);
+            if (\function_exists($f)) {
+                $data = \call_user_func_array($f, $params);
             }
         } else {
             if (method_exists($o, $f)) {
-                $data = call_user_func_array(array($o, $f), $params);
+                $data = \call_user_func_array(array($o, $f), $params);
             }
         }
 
         // If function returns a value,
-        if (isset($data)) {
-            if (is_array($data) && isset($field[$property]) && is_array($field[$property])) {
+        if (null !== $data) {
+            if (\is_array($data) && isset($field[$property]) && \is_array($field[$property])) {
                 // Combine field and @data-field together.
                 $field[$property] += $data;
             } else {