+++ File/src/File.php
@@ -58,7 +58,7 @@ class File implements FileInterface
      */
     public static function instance($filename)
     {
-        if (!is_string($filename) && $filename) {
+        if (!\is_string($filename) && $filename) {
             throw new \InvalidArgumentException('Filename should be non-empty string');
         }
         if (!isset(static::$instances[$filename])) {
@@ -188,7 +188,7 @@ class File implements FileInterface
     public function lock($block = true)
     {
         if (!$this->handle) {
-            if (!$this->mkdir(dirname($this->filename))) {
+            if (!$this->mkdir(\dirname($this->filename))) {
                 throw new \RuntimeException('Creating directory failed for ' . $this->filename);
             }
             $this->handle = @fopen($this->filename, 'cb+');
@@ -199,7 +199,11 @@ class File implements FileInterface
             }
         }
         $lock = $block ? LOCK_EX : LOCK_EX | LOCK_NB;
-        return $this->locked = $this->handle ? flock($this->handle, $lock) : false;
+
+        // Some filesystems do not support file locks, only fail if another process holds the lock.
+        $this->locked = flock($this->handle, $lock, $wouldblock) || !$wouldblock;
+
+        return $this->locked;
     }
 
     /**
@@ -239,7 +243,7 @@ class File implements FileInterface
      */
     public function writable()
     {
-        return is_writable($this->filename) || $this->writableDir(dirname($this->filename));
+        return is_writable($this->filename) || $this->writableDir(\dirname($this->filename));
     }
 
     /**
@@ -268,7 +272,7 @@ class File implements FileInterface
             $this->content = null;
         }
 
-        if (!is_string($this->raw)) {
+        if (!\is_string($this->raw)) {
             $this->raw = $this->load();
         }
 
@@ -314,26 +318,34 @@ class File implements FileInterface
             $this->content($data);
         }
 
-        if (!$this->locked) {
-            // Obtain blocking lock or fail.
-            if (!$this->lock()) {
-                throw new \RuntimeException('Obtaining write lock failed on file: ' . $this->filename);
+        $filename = $this->filename;
+        $dir = \dirname($filename);
+
+        if ($this->handle) {
+            $tmp = true;
+            // As we are using non-truncating locking, make sure that the file is empty before writing.
+            if (@ftruncate($this->handle, 0) === false || @fwrite($this->handle, $this->raw()) === false) {
+                // Writing file failed, throw an error.
+                $tmp = false;
             }
-            $lock = true;
+        } else {
+            // Create file with a temporary name and rename it to make the save action atomic.
+            $tmp = tempnam($dir, basename($filename));
+            if ($tmp && @file_put_contents($tmp, $this->raw()) && @rename($tmp, $filename)) {
+                // We need to chmod() the file as it's using 0600 permissions.
+                @chmod($filename, 0666 & ~umask());
+            } elseif ($tmp) {
+                @unlink($tmp);
+                $tmp = false;
+            }
         }
 
-        // As we are using non-truncating locking, make sure that the file is empty before writing.
-        if (@ftruncate($this->handle, 0) === false || @fwrite($this->handle, $this->raw()) === false) {
-            $this->unlock();
-            throw new \RuntimeException('Saving file failed: ' . $this->filename);
-        }
-
-        if (isset($lock)) {
-            $this->unlock();
+        if ($tmp === false) {
+            throw new \RuntimeException('Failed to save file ' . $this->filename);
         }
 
         // Touch the directory as well, thus marking it modified.
-        @touch(dirname($this->filename));
+        @touch($dir);
     }
 
     /**
@@ -407,20 +419,21 @@ class File implements FileInterface
 
     /**
      * @param  string  $dir
-     * @return bool
-     * @throws \RuntimeException
-     * @internal
      */
-    protected function mkdir($dir)
+    private function mkdir($dir)
     {
         // Silence error for open_basedir; should fail in mkdir instead.
-        if (!@is_dir($dir)) {
-            $success = @mkdir($dir, 0777, true);
+        if (@is_dir($dir)) {
+            return true;
+        }
 
-            if (!$success) {
-                $error = error_get_last();
+        $success = @mkdir($dir, 0777, true);
 
-                throw new \RuntimeException("Creating directory '{$dir}' failed on error {$error['message']}");
+        if (!$success) {
+            // Take yet another look, make sure that the folder doesn't exist.
+            clearstatcache(true, $dir);
+            if (!@is_dir($dir)) {
+                return false;
             }
         }
 
@@ -435,7 +448,7 @@ class File implements FileInterface
     protected function writableDir($dir)
     {
         if ($dir && !file_exists($dir)) {
-            return $this->writableDir(dirname($dir));
+            return $this->writableDir(\dirname($dir));
         }
 
         return $dir && is_dir($dir) && is_writable($dir);