+++ File/src/MoFile.php
@@ -64,7 +64,7 @@ class MoFile extends File
     {
         $this->endian = 'V';
         $this->str = $var;
-        $this->len = strlen($var);
+        $this->len = \strlen($var);
 
         $magic = $this->readInt() & 0xffffffff;