+++ ResourceLocator/src/UniformResourceLocator.php
@@ -100,9 +100,9 @@ class UniformResourceLocator implements ResourceLocatorInterface
     {
         $list = [];
         foreach((array) $paths as $path) {
-            if (is_array($path)) {
+            if (\is_array($path)) {
                 // Support stream lookup in ['theme', 'path/to'] format.
-                if (count($path) !== 2 || !is_string($path[0]) || !is_string($path[1])) {
+                if (\count($path) !== 2 || !\is_string($path[0]) || !\is_string($path[1])) {
                     throw new \BadMethodCallException('Invalid stream path given.');
                 }
                 $list[] = $path;
@@ -127,7 +127,7 @@ class UniformResourceLocator implements ResourceLocatorInterface
             if (!$override || $override == 1) {
                 $list = $override ? array_merge($paths, $list) : array_merge($list, $paths);
             } else {
-                $location = array_search($override, $paths) ?: count($paths);
+                $location = array_search($override, $paths, true) ?: \count($paths);
                 array_splice($paths, $location, 0, $list);
                 $list = $paths;
             }
@@ -191,7 +191,7 @@ class UniformResourceLocator implements ResourceLocatorInterface
      */
     public function __invoke($uri)
     {
-        if (!is_string($uri)) {
+        if (!\is_string($uri)) {
             throw new \BadMethodCallException('Invalid parameter $uri.');
         }
         return $this->findCached($uri, false, true, false);
@@ -228,7 +228,7 @@ class UniformResourceLocator implements ResourceLocatorInterface
      */
     public function normalize($uri, $throwException = false, $splitStream = false)
     {
-        if (!is_string($uri)) {
+        if (!\is_string($uri)) {
             if ($throwException) {
                 throw new \BadMethodCallException('Invalid parameter $uri.');
             }
@@ -284,7 +284,7 @@ class UniformResourceLocator implements ResourceLocatorInterface
      */
     public function findResource($uri, $absolute = true, $first = false)
     {
-        if (!is_string($uri)) {
+        if (!\is_string($uri)) {
             throw new \BadMethodCallException('Invalid parameter $uri.');
         }
         return $this->findCached($uri, false, $absolute, $first);
@@ -301,7 +301,7 @@ class UniformResourceLocator implements ResourceLocatorInterface
      */
     public function findResources($uri, $absolute = true, $all = false)
     {
-        if (!is_string($uri)) {
+        if (!\is_string($uri)) {
             throw new \BadMethodCallException('Invalid parameter $uri.');
         }
 
@@ -442,10 +442,10 @@ class UniformResourceLocator implements ResourceLocatorInterface
             }
 
             // Remove prefix from filename.
-            $filename = '/' . trim(substr($file, strlen($prefix)), '\/');
+            $filename = '/' . trim(substr($file, \strlen($prefix)), '\/');
 
             foreach ($paths as $path) {
-                if (is_array($path)) {
+                if (\is_array($path)) {
                     // Handle scheme lookup.
                     $relPath = trim($path[1] . $filename, '/');
                     $found = $this->find($path[0], $relPath, $array, $absolute, $all);